Modularized the View Individual Record page into a separate file, and added buttons to view individual ticket by /view_ticket/id
