+++ templates/view_ticket.html
@@ -33,5 +33,6 @@
         <td>{{ ticket.updated_at }}</td>
       </tr>
     </table>
+    <a href="{{ url_for('view_tickets.view_tickets') }}">Back to Tickets</a>
   </body>
 </html>
