+++ templates/view_tickets.html
@@ -13,11 +13,9 @@
         <tr>
           <th>ID</th>
           <th>Title</th>
-          <th>Description</th>
           <th>Status</th>
           <th>Priority</th>
-          <th>Created At</th>
-          <!-- <th>Comments</th> -->
+          <th>Details</th>
         </tr>
       </thead>
       <tbody>
@@ -25,11 +23,16 @@
         <tr>
           <td>{{ ticket.id }}</td>
           <td>{{ ticket.title }}</td>
-          <td>{{ ticket.description }}</td>
           <td>{{ ticket.status }}</td>
           <td>{{ ticket.priority }}</td>
-          <td>{{ ticket.created_at }}</td>
-          <!-- <td><button>View Comments</button></td> -->
+          <td>
+            <a
+              href="{{ url_for('view_individual_ticket.view_ticket', ticket_id=ticket.id) }}"
+              class="submit"
+            >
+              View More Details
+            </a>
+          </td>
         </tr>
         {% endfor %}
       </tbody>
@@ -37,5 +40,6 @@
     {% else %}
     <p>You have not submitted any tickets.</p>
     {% endif %}
+    <a href="{{ url_for('home.home_screen') }}">Back to Home</a>
   </body>
 </html>
